add best configuration
