+++ .vscode/launch.json
@@ -1,0 +1,26 @@
+// {
+//     "configurations": [
+//         {
+//             "name": "Launch file",
+//             "type": "go",
+//             "request": "launch",
+//             "mode": "debug",
+//             "program": "main.go"
+//         }
+//     ]
+// }
+
+{
+    "version": "0.2.0",
+    "configurations": [
+      {
+        "name": "Go: Launch",
+        "type": "go",
+        "request": "launch",
+        "program": "${workspaceFolder}/main.go",
+        "args": [],
+        "mode": "debug",
+
+      }
+    ]
+  }
