+++ internal/repository/dbrepo/mysql.go
@@ -1,4 +1,4 @@
-package repository
+package dbrepo
 
 import (
 	"context"
@@ -14,11 +14,11 @@ func NewMySQL(db *sql.DB) *MySQL {
 	return &MySQL{db: db}
 }
 
-func (m *MySQL) CreateUser(ctx context.Context, user *structs.User) (int64, error) {
-	// Implement actual create user logic here
-	// For now, this is just a stub
-	return 1, nil // Replace with actual ID generation
-}
+// func (m *MySQL) CreateUser(ctx context.Context, user *structs.User) (int64, error) {
+// 	// Implement actual create user logic here
+// 	// For now, this is just a stub
+// 	return 1, nil // Replace with actual ID generation
+// }
 
 func (m *MySQL) GetUser(ctx context.Context, id int64) (*structs.User, error) {
 	// Implement actual get user logic here
@@ -36,14 +36,14 @@ func (m *MySQL) GetUser(ctx context.Context, id int64) (*structs.User, error) {
 	}, nil
 }
 
-func (m *MySQL) UpdateUser(ctx context.Context, user *structs.User) error {
-	// Implement actual update user logic here
-	// For now, this is just a stub
-	return nil
-}
+// func (m *MySQL) UpdateUser(ctx context.Context, user *structs.User) error {
+// 	// Implement actual update user logic here
+// 	// For now, this is just a stub
+// 	return nil
+// }
 
-func (m *MySQL) DeleteUser(ctx context.Context, id int64) error {
-	// Implement actual delete user logic here
-	// For now, this is just a stub
-	return nil
-}
+// func (m *MySQL) DeleteUser(ctx context.Context, id int64) error {
+// 	// Implement actual delete user logic here
+// 	// For now, this is just a stub
+// 	return nil
+// }
