+++ internal/repository/dbrepo/postgres.go
@@ -1,0 +1,46 @@
+package dbrepo
+
+import (
+	"database/sql"
+	"kissandeat/internal/structs"
+)
+
+type Postgres struct {
+	Db *sql.DB
+}
+
+func NewPostgres(db *sql.DB) *Postgres {
+	return &Postgres{Db: db}
+}
+
+//	func (r *Postgres) CreateUser(ctx context.Context, user *structs.User) (int64, error) {
+//		// Implement actual create user logic here
+//		// For now, this is just a stub
+//		return 1, nil // Replace with actual ID generation
+//	}
+func (r *Postgres) GetUser(id int64) (*structs.User, error) {
+	// Implement actual get user logic here
+	// For now, this is just a stub
+	return &structs.User{
+		ID:        5,
+		Username:  "Stubbed User",
+		Password:  "stubbed_password", // Replace with a secure hash or placeholder
+		Email:     "stubbed@example.com",
+		FirstName: "John",
+		LastName:  "Doe",
+		Gender:    "male",
+		Role:      "husband",
+		FamilyID:  1,
+	}, nil
+}
+
+// func (r *Postgres) UpdateUser(ctx context.Context, user *structs.User) error {
+// 	// Implement actual update user logic here
+// 	// For now, this is just a stub
+// 	return nil
+// }
+// func (r *Postgres) DeleteUser(ctx context.Context, id int64) error {
+// 	// Implement actual delete user logic here
+// 	// For now, this is just a stub
+// 	return nil
+// }
