+++ internal/service/dbservice/interfaces.go
@@ -1,0 +1,1 @@
+package dbservice
