+++ internal/service/dbservice/service.go
@@ -1,0 +1,1 @@
+package dbservice
