+++ internal/service/dbservice/storage.go
@@ -1,0 +1,1 @@
+package dbservice
